Add support for paths to the controller while calling the rest or the service interface.
Previously, the service or restService can only be accessed if the FooService class
is located in a directory underneath the PartKeepr directory. Using namespaces
was not possible, so an efficient grouping is not possible. Now the controller
can be called with a separation of '.' which is then computed to the complete
path. Now it is possible to access services also using nested namespaces.

For eg. The old style was: FooBar1, FooBar2, .... which was mapped to
class PathKeepr/FooBar1/FooBar1Service
Now it is possible to use: Foo.Bar1, Foo.Bar2 on controller side, which is
then mapped to PathKeepr/Foo/Bar1/Bar1Service.