Adds label printing to the project.
For printing labels, the library TCPDF is necessary. We add a first step towards
have the possibility to use labels which are located on a bigger sheet. We
implemented a plugin structure, which can be easily extended by a user to
have a custom renderer for its own style. We provide a default renderer which
renders to a PDF document. This renderer can also be parametrized from the
user and should extended to the needs of the user as they are common.