Adds a trigger_error if the TCPDF define K_TCPDF_CALLS_IN_HTML is set to true to ensure nobody will open a security hole.
This flag enables the usage of an extended version with inlining function
calls to php (which will be evaled) directly (inside a html text write).
As long as we use only code from our class, this is no problem, but an
implementer of a plugin may not think of this risk and should be reminded
with this.