Implementing a way to support notification / listener for a client.
The client can use the EventNotificationService to register to one or more
events and can then use a long polling (or short one) to check if an event
was present. The client can than react on this event.
This has the advantage, especially if a long poll is used, of a lesser
server load in these cases with a decreased notification time.