Fixes a null object exception on FootPrintComboBox.
The listeners are not automatically removed from the store and since the
store has a much longer lifetime, the event is fired and tries to call
an dangling reference to this class.

There are more such things pending in the project due to code duplication
artifacts which should be omitted.