Fixes picture upload if open_basedir is in effect, with excluded system-wide temp. dir
The patch makes use of the upload_tmp_dir if set, and falls back to sys_get_temp_dir.
While sys_get_temp_dir is using the environment variable TMPDIR to override the default,
upload_tmp_dir can be set in: php.ini, apache's VirtualHost with php_*_value or in the .htaccess.