Don't expose the project property, fixes #579
The root cause was that circular references are serialized as ID. Example:

Retrieve ```api/projects/25```. ```project``` propeties of any `ProjectPart` are now referenced as ID and not as object, as such the deserializer doesn't fill the object properly.

The proper solution is not to include a circular reference when retrieving projects.