Implement #341 / #404
Implements the feature described in issue #341 / #404 (related / duplicates)

This simply adds a check to the report calculation which takes the packagingUnit into account.

If the part count (so the missing parts which have to be ordered) is not larger than the packagingUnit of the part distributor, that distributor is omitted (by the continue keyword).